# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Doc Api",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/categories": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Список",
                "operationId": "ac9c4d978ea4e5b72915ca9728e5b8b1",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Фен для волос"
                                                    },
                                                    "slug": {
                                                        "type": "string",
                                                        "example": "fen-dlya-volos"
                                                    },
                                                    "external_id": {
                                                        "type": "integer",
                                                        "example": 3241
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Category"
                ],
                "summary": "Создание",
                "operationId": "cca49a87d9b0ebb9a98ad7c154b71737",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "title": {
                                                "type": "string",
                                                "example": "Фен для волос"
                                            },
                                            "slug": {
                                                "type": "string",
                                                "example": "fen-dlya-volos"
                                            },
                                            "external_id": {
                                                "type": "integer",
                                                "example": 3241
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Фен для волос"
                                                },
                                                "slug": {
                                                    "type": "string",
                                                    "example": "fen-dlya-volos"
                                                },
                                                "external_id": {
                                                    "type": "integer",
                                                    "example": 3241
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/categories/{slug}": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Единичная запись",
                "operationId": "015a7b9d140957c1a14c67c5318977e2",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Slug Категории",
                        "required": true,
                        "example": "fen-dlya-volos"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                }
            },
            "put": {
                "tags": [
                    "Category"
                ],
                "summary": "Обновление",
                "operationId": "88024c0076e21d876d2e023d0ffe2901",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Slug Категории",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "string": {
                                "summary": "An int value.",
                                "value": "1"
                            }
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "title": {
                                                "type": "string",
                                                "example": "Фен для волос изменить"
                                            },
                                            "external_id": {
                                                "type": "integer",
                                                "example": 3241
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Фен для волос"
                                                },
                                                "slug": {
                                                    "type": "string",
                                                    "example": "fen-dlya-volos"
                                                },
                                                "external_id": {
                                                    "type": "integer",
                                                    "example": 3241
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Category"
                ],
                "summary": "Удаление",
                "operationId": "ec1f85af84113d27b3790691aee59748",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Slug Категории",
                        "required": true,
                        "example": "fen-dlya-volos"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "done"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}